+++ src/components/cart/cartHeader/CartHeader.tsx
@@ -6,8 +6,8 @@ import {
   cartSelectedState,
 } from '@/recoil/atoms/cartState';
 
-import CartHeaderButton from './CartHeaderButton';
 import DeleteSelectedButton from './DeleteSelectedButton';
+import CartHeaderButton from './CartHeaderButton';
 
 const CartHeader = () => {
   const [selectedCartList, setSelectedCartList] =
