+++ src/components/cart/index.ts
@@ -1,5 +1,5 @@
 import CartFooter from './CartFooter';
-import CartHeader from './CartHeader';
+import CartHeader from './cartHeader/CartHeader';
 import CartItem from './cartItem/CartItem';
 import EmptyCartItem from './cartItem/EmptyCartItem';
 import CartNotice from './CartNotice';
